+++ tiledb/serde/src/lib.rs
@@ -5,6 +5,3 @@ extern crate tiledb_sys as ffi;
 
 pub mod array;
 pub mod filter;
-
-#[cfg(any(test, feature = "proptest-strategies"))]
-pub mod strategy;
